requirements.txt
Flask
Flask-WTF
replicate
openai
tiktoken
langchain
python-dotenv
